Added analytics

+++ OmniKDS/Views/OPKitchenItemView.swift
@@ -104,7 +104,7 @@ class OPKitchenItemView: UICollectionViewCell, UITableViewDelegate, UITableViewD
         order?.isOpen = false
         order?.closedAt = Date()
         let seconds = KSDateUtil.getTimeDiffBetweenDate1(order?.orderDate!, andDate2: order?.closedAt!)
-        let mins = Float(seconds/60)
+        let mins = Float(Float(seconds)/60) as Float
         order?.processingTime = mins
         let max = NSNumber(value: sharedKitchen!.turnToRedAfter)
